Set-up webpack and babel with firebase
Webpack and babel have been set-up to export main-[hash].js with a
main.[hash].css file. webpack.config.js has been split into a common,
dev, and prod version. Static files in the /src/assets folder are copied
to a static folder in the dist directory. Tested to be working on
firebase with `firebase deploy`.